+++ src/error_handling/error_messages.rs
@@ -25,6 +25,7 @@ pub const SEMANTIC_MISMATCHED_TYPES: u32 = 21;
 pub const SEMANTIC_DUPLICATED_REFERENCE: u32 = 22;
 pub const SEMANTIC_MISMATCHED_ARGUMENT_COUNT: u32 = 23;
 pub const SYNTAX_INVALID_ARRAY_ACCESS: u32 = 24;
+pub const SEMANTIC_IMMUTABLE_VARIABLE: u32 = 25;
 
 /// Reads the error codes from the error code list
 pub fn error_code_to_string() -> String {