+++ src/semantic/checks/top_level_declaration.rs
@@ -25,9 +25,7 @@ impl SemanticCheck for Statement<'_> {
             Statement::Conditional(c) => c.check_semantics(scope),
             Statement::ForLoop(f) => f.check_semantics(scope),
             Statement::WhileLoop(w) => w.check_semantics(scope),
-            Statement::Assignment(_assignment) => {
-                unimplemented!("Semantic check for an assignment")
-            }
+            Statement::Assignment(a) => a.check_semantics(scope),
         }
     }
 }